ubusd_acl: use size_t for strlen result in ubusd_acl_alloc_obj
len held the result of strlen() in an int. It is later passed as
len + 1 to calloc_a(), whose variadic length arguments are read via
va_arg as size_t in __calloc_a(). Passing an int where size_t is
expected is undefined behaviour and can yield a wrong allocation
size on LP64 platforms.

Link: https://github.com/openwrt/ubus/pull/20

+++ ubusd_acl.c
@@ -321,7 +321,7 @@ static struct ubusd_acl_obj*
 ubusd_acl_alloc_obj(struct ubusd_acl_file *file, const char *obj)
 {
 	struct ubusd_acl_obj *o;
-	int len = strlen(obj);
+	size_t len = strlen(obj);
 	char *k;
 	bool partial = false;
